// src/lib/api/generated/model/campaignStatusEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `DRAFT` - Draft
* `SCHEDULED` - Scheduled
* `RUNNING` - Running
* `PAUSED` - Paused
* `COMPLETED` - Completed
* `CANCELED` - Canceled
 */
export type CampaignStatusEnum = typeof CampaignStatusEnum[keyof typeof CampaignStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CampaignStatusEnum = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  RUNNING: 'RUNNING',
  PAUSED: 'PAUSED',
  COMPLETED: 'COMPLETED',
  CANCELED: 'CANCELED',
} as const;