// src/lib/api/generated/model/deliveryChannelEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `EMAIL` - Email
* `WEBHOOK` - Webhook
 */
export type DeliveryChannelEnum = typeof DeliveryChannelEnum[keyof typeof DeliveryChannelEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeliveryChannelEnum = {
  EMAIL: 'EMAIL',
  WEBHOOK: 'WEBHOOK',
} as const;