// src/lib/api/generated/model/sendConditionEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ALWAYS` - Always
* `IF_NOT_RESPONDED` - If Not Responded
 */
export type SendConditionEnum = typeof SendConditionEnum[keyof typeof SendConditionEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SendConditionEnum = {
  ALWAYS: 'ALWAYS',
  IF_NOT_RESPONDED: 'IF_NOT_RESPONDED',
} as const;