// src/lib/api/generated/model/stepTypeEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `INITIAL` - Initial
* `REMINDER` - Reminder
* `FOLLOWUP` - Follow-up
 */
export type StepTypeEnum = typeof StepTypeEnum[keyof typeof StepTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StepTypeEnum = {
  INITIAL: 'INITIAL',
  REMINDER: 'REMINDER',
  FOLLOWUP: 'FOLLOWUP',
} as const;